Add configurable styles
Added the ability to define sections within the watson config to change
the color of any element. You can define both the background and
foreground color of the element.

A third option is available if
enabled in the general settings called function. Function takes in the
name of a function defined in file found at
~/.config/watson/functions.py. This file is given the content of the
element and should return a parsed version. This is useful to change
how tags or id's are displayed. This functionality is only enabled
when options.load_functions_file is set to true.